NAVI-3016 - Добавлено получение карты - Добавлено получение прогресса при запросе карты

+++ shared/src/commonMain/kotlin/auto/atom/yandexmapdownloadmanager/map/OfflineRegion.kt
@@ -13,5 +13,5 @@ data class OfflineRegion(
     val size: Long,
     val children: MutableList<OfflineRegion> = mutableListOf(),
     val state: RegionState = RegionState.NOT_DOWNLOADED,
-    val downloadProgress: Int? = null,
+    val downloadProgress: Float? = null,
 )
